Add boolean dtrRtsOn parameter to CreateUsbSerialDevice.  This allows the user to decide if the lines are raised at init or not.   Default is that RTS and DTR will be high so as to not break previous functionality.
Not all devices use RTS and DTR for flow control, Arduinos are one example of this.  Depending on the Arduino, they use one of those lines to reset the device.  Its common for DTR and RTS to be high on a Serial/Com port when connected so the device always resets.

Other serial devices use the DTR line for power (and possibly RTS as well).  If the line is high when connected, the device will power on when it may not be desired.

Note: This is only implemented for FTDI, Pl2303, and CH34x devices.  I haven't tested it yet, but its a fairly trivial change and should pose no problem